+++ apps/tako/src/events/error.ts
@@ -5,6 +5,6 @@ import type { Event } from './index.ts';
 export default {
 	name: Events.Error,
 	async execute(error) {
-		logger.error(error);
+		logger.error(error.message);
 	},
 } satisfies Event<'error'>;